RUS: Initial version of rus
[R]un [U]ntil [S]uccess

rus is a simple bash function that allows a user to repeatedly
run a command until it succeeds.

To use this command, simply prepend "rus" to the beginning of any
command you wish to run until a return code of "0" is received.

i.e. the command you wish to run should be passed as arguments to
rus.

Ex: rus cat /tmp/myfile

This example will attempt to cat /tmp/myfile until successful.
Once successful, the result will be output, followed by a count of
required attempts.